Refactored Data Access Layer

+++ SLIP-WebServer/src/main/java/dataAccessLayer/SessionPayloadDAO.java
@@ -7,18 +7,12 @@ import model.ServerPayload;
 import model.Session;
 public interface SessionPayloadDAO {
   
-//  /**
-//   * Set the datasource the JDBC template will use
-//   * @param dataSource
-//   */
-//  public void setDataSource(DataSource dataSource);
-  
   /**
    * Return all the payloads that were received in a given session ID
    * @param sessionID
    * @return
    */
-  public List<ServerPayload> getAllPayloads(int sessionID);
+  public List<ServerPayload> getAllPayloads(long sessionID);
   
   /**
    * Return a specific payload  given a session ID
@@ -26,7 +20,7 @@ public interface SessionPayloadDAO {
    * @param payloadID The ID of the payload to be retrieved
    * @return
    */
-  public ServerPayload getPayload(int sessionID, int payloadID);
+  public ServerPayload getPayload(long sessionID, long payloadID);
   
   /**
    * Insert a frame of payloads into the DB
@@ -39,7 +33,7 @@ public interface SessionPayloadDAO {
    * @param sessionID The session ID of the payload to be inserted
    * @param payload
    */
-  public void insertPayload(int sessionID, ServerPayload payload);
+  public void insertPayload(long sessionID, ServerPayload payload);
   
   /**
    * Return all the payloads that arrived later than the given timestamp
